short term fix for te linear gemms (#126)
The user's profile uses Transformer engine ops like _Linear and
_LayerNormLinear. These ops directly launch the gemm kernel in the
profile without a gemm op.
As a temporary fix we insert synthetic gemm ops based on shapes parsed
from the higher-level TE ops. This allows us to use the same perf
metrics flow as usual.
This is to unblock the user but we need to think of a longer term
solution as well - for future work

---------