Add script to filter trace to range of given user annotation (#163)
A common pattern is to use the PyTorch profiler like so:
```
x = torch.randn((1, 1), requires_grad=True)
with torch.autograd.profiler.profile() as prof:
    y = x ** 2
    with torch.autograd.profiler.record_function("label-z"): # label the block
        z = y ** 3
```
The labels can indicate regions of interest, e.g., particular stages of
model inference or training. It is often interesting to analyze such
labeled sections independently rather than in the context of the entire
profile trace.

This PR provides a standalone script where a user can filter their trace
file first to a specific label and then proceed with TraceLens analysis
on the filtered file instead.

A related feature was requested in #90 to add analysis of labeled
regions to TraceLens. This PR does _not_ provide that functionality, as
it seems to require quite a bit more work and modifying the internals.