Add roofline_analyzer 0.1.1 (#182)
I've created a python cli application that 
as an input takes the XLSX that is, to my knowledge. generated by
`generate_perf_report.py` from
https://github.com/AMD-AIG-AIMA/TraceLens/pull/138

and when combined with a config-file that has e.g. MI300X bfloat16
values:

![image](https://github.com/user-attachments/assets/e8ab7f6e-fa19-4537-aa35-ffd3e1a43eb5)

It will generate a naive roofline analysis for the GEMMs used in e.g. a
client model:

![image](https://github.com/user-attachments/assets/06e3a4b1-fc6b-43af-86fe-70bfd2fde24f)

![image](https://github.com/user-attachments/assets/3fb5c4f2-bfd5-40c0-96a9-c8524d523c2e)

and it will generate a new XLSX as an output, that has:
1. **bound_type**: what is the first limiting roofline (memory or
compute), based on the "AI ridge point"
2. **reference_roofline**: what is used as the reference value for
roofline distance calculation
3. **bound_distance**: what is the distance from kernel to the
maf_roofline in question
4. **bound_distance_pct**: above in percentage


![image](https://github.com/user-attachments/assets/d6e4087b-b499-4fde-8a7d-a3465c75c68b)
and AI ridge points for both the max achievable flops (maf) and max
theoretical flops (mtf) i.e. the point that divides whether the kernel
is memory or compute limited:

![image](https://github.com/user-attachments/assets/a32c7b9d-1abe-4246-9b9c-417d1308383b)

What is new in 0.1.1 compared to 0.1.0 is that the kernels will get a
rolling numbering, which is also used to indicate which point on the
plot belongs to which kernel:

![image](https://github.com/user-attachments/assets/2ab46e4c-2363-43ac-8f8d-48f2a094f159)

![image](https://github.com/user-attachments/assets/835baeb7-5fc9-4836-85a1-3c012022f0f9)

For additional information, consult the extensive README.md