Fix/jax analyses do not parse file header (#218)
When profiling a jit-compiled function like the train_step in Maxtext,
the parsed xplane.pb will have a header line starting with "HloModule ".
This line has no information that we currently want to parse but at the
same time, it causes an exception to be raised. This commit completely
ignores the header line.

---------

Co-authored-by: Ossi Lehtikangas <ossi.lehtikangas@amd.com>