[NcclAnalyser]: Add option to warn instead of raise on metadata consistency failure (#271)
We have an analysis/report generation workflow where occasionally we
want to override the default behavior of `NcclAnalyser` and allow it to
proceed regardless of possible metadata mismatches.

To support this workflow, this PR adds a new boolean argument
`strict_metadata_check` to the `NcclAnalyser` interface. With the
default value, the current behavior is unchanged and the methods will
still raise a `ValueError` if a metadata mismatch is encounterd. If user
explicitly sets `strict_metadata_check=False`, a warning is emitted
instead of raising an error.

This is part of effort to clean up #138 and prepare it for merging.