feat: added ability to augment perf report via extensions (#446)
Added a feature to the Perf Report Generation Script (in Pytorch) that
allows any Perf Report Extensions to augment the base perf report with
extra columns. This change allows for custom dataframes to be added to
TraceLens Perf Report.

P.S: This commit also cleans up the package import order in the perf
report script; thought I would leave this change in for beautification